requirements/default.txt
numpy==1.16.4
biopython==1.73.0
matplotlib==3.1.1
bokeh==0.13.0
pandas==1.0.3
seaborn==0.10.0
